Extract settings page content rendering
Generated-By: PostHog Code
Task-Id: 5579dbc2-d5dd-440f-8fb5-3cbd00ed0957
